# api/openapi.yaml
openapi: 3.1.0
info:
  title: List API
  version: 0.1.0
  description: |
    Contract-first API (ADR-0002). This file is the source of truth: change it
    before changing handlers or client code. CI rejects breaking changes.
    All errors are RFC 7807 `application/problem+json`.
servers:
  - url: /
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/tasks/bulk:
    post:
      tags: [tasks]
      operationId: bulkCreateTasks
      summary: Quick capture - create many tasks at once, all or nothing (E10-S1, ADR-0024)
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
        - name: Idempotency-Key
          in: header
          required: true
          description: 8-200 characters. A retry with the same key and body returns the first response.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BulkCreateTasksRequest" }
      responses:
        "201":
          description: Created (or replayed)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskList" }
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/tasks:
    get:
      tags: [tasks]
      operationId: listTasks
      summary: Tasks in the active workspace (filtered; no pagination - see A2)
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/TaskStatus" } }
        - { name: mine, in: query, description: "true = assigned to me (overrides assigneeId)", schema: { type: string, enum: ["true"] } }
        - { name: assigneeId, in: query, schema: { type: string } }
        - { name: labelId, in: query, schema: { type: string } }
        - { name: clientId, in: query, schema: { type: string } }
        - { name: parentId, in: query, description: Only subtasks of this task, schema: { type: string } }
      responses:
        "200":
          description: Tasks, due date first (no date last), then newest
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskList" }
        default:
          $ref: "#/components/responses/Problem"
    post:
      tags: [tasks]
      operationId: createTask
      summary: Create a task (caller becomes its creator)
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTaskRequest" }
      responses:
        "201":
          description: Created
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskResponse" }
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/tasks/{taskId}:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    get:
      tags: [tasks]
      operationId: getTask
      summary: One task with the caller's permissions
      responses:
        "200":
          description: Task
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskResponse" }
        default:
          $ref: "#/components/responses/Problem"
    patch:
      tags: [tasks]
      operationId: updateTask
      summary: Change fields (only those present). Optimistic concurrency via If-Match.
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTaskRequest" }
      responses:
        "200":
          description: Updated
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskResponse" }
        default:
          $ref: "#/components/responses/Problem"
    delete:
      tags: [tasks]
      operationId: deleteTask
      summary: Delete a task
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "204":
          description: Deleted
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/tasks/{taskId}/assignees:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    put:
      tags: [tasks]
      operationId: setTaskAssignees
      summary: Replace assignees (workspace members only)
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserIdsRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskResponse" }
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/tasks/{taskId}/owners:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    put:
      tags: [tasks]
      operationId: setTaskOwners
      summary: Replace owners (creator only; the creator is never listed as an owner)
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserIdsRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskResponse" }
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/tasks/{taskId}/labels:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    put:
      tags: [tasks]
      operationId: setTaskLabels
      summary: Replace labels (labels of this workspace only)
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LabelIdsRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskResponse" }
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
  headers:
    ETag:
      description: 'Task version, e.g. "3". Send it back in If-Match.'
      schema: { type: string }
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: string }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: The ETag you edited from. A stale value returns 409 conflict; a missing one 428.
      schema: { type: string }
    CsrfHeader:
      name: X-Requested-With
      in: header
      required: true
      description: Must be "app" on every unsafe request (CSRF defence, ADR-0008).
      schema:
        type: string
        enum: [app]
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    TaskSource:
      type: string
      enum: [MEETING_NOTE, PERSONAL]
    BulkCreateTasksRequest:
      type: object
      required: [source, startDate, endDate, titles]
      properties:
        source: { $ref: "#/components/schemas/TaskSource" }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        titles:
          type: array
          minItems: 1
          maxItems: 100
          items: { type: string }
...
    TaskStatus:
      type: string
      enum: [BACKLOG, TODO, IN_PROGRESS, TESTING, DONE, CANCELED]
    TaskPriority:
      type: string
      enum: [URGENT, HIGH, MEDIUM, LOW, NONE]
    LabelColor:
      type: string
      enum: [gray, red, orange, yellow, green, teal, blue, purple, pink]
    PersonRef:
      type: object
      required: [id, name, image]
      properties:
        id: { type: string }
        name: { type: string }
        image: { type: [string, "null"] }
    Label:
      type: object
      required: [id, name, color]
      properties:
        id: { type: string }
        name: { type: string }
        color: { $ref: "#/components/schemas/LabelColor" }
    TaskViewer:
      type: object
      description: The caller's permissions on this task (E5-S1).
      required: [canManage, canManageOwners, isAssignee, canRequest]
      properties:
        canManage: { type: boolean }
        canManageOwners: { type: boolean }
        isAssignee: { type: boolean }
        canRequest: { type: boolean, description: "An assignee who can't manage: proposes changes as requests (E5-S2)" }
    TaskSummary:
      type: object
      required: [id, title, status, priority, startDate, endDate, dueDate, assignees, labels, createdBy, version, updatedAt, client, parent, counts, source, viewer]
      properties:
        id: { type: string }
        title: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        priority: { $ref: "#/components/schemas/TaskPriority" }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        dueDate: { type: [string, "null"], format: date }
        assignees:
          type: array
          items: { $ref: "#/components/schemas/PersonRef" }
        labels:
          type: array
          items: { $ref: "#/components/schemas/Label" }
        createdBy: { $ref: "#/components/schemas/PersonRef" }
        version: { type: integer }
        updatedAt: { type: string, format: date-time }
        client:
          anyOf:
            - { $ref: "#/components/schemas/ClientRef" }
            - { type: "null" }
        parent:
          anyOf:
            - { $ref: "#/components/schemas/TaskRef" }
            - { type: "null" }
        counts: { $ref: "#/components/schemas/TaskCounts" }
        source:
          description: Set on tasks made by quick capture
          anyOf:
            - { $ref: "#/components/schemas/TaskSource" }
            - { type: "null" }
        viewer: { $ref: "#/components/schemas/TaskViewer" }
    Task:
      type: object
      required: [id, title, description, status, priority, startDate, endDate, dueDate, assignees, owners, labels, createdBy, createdAt, updatedAt, version, viewer, client, parent, subtasks, checklist, comments, attachments, requests]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/TaskStatus" }
        priority: { $ref: "#/components/schemas/TaskPriority" }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        dueDate: { type: [string, "null"], format: date }
        assignees:
          type: array
          items: { $ref: "#/components/schemas/PersonRef" }
        owners:
          type: array
          items: { $ref: "#/components/schemas/PersonRef" }
        labels:
          type: array
          items: { $ref: "#/components/schemas/Label" }
        createdBy: { $ref: "#/components/schemas/PersonRef" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer }
        viewer: { $ref: "#/components/schemas/TaskViewer" }
        client:
          anyOf:
            - { $ref: "#/components/schemas/ClientRef" }
            - { type: "null" }
        parent:
          anyOf:
            - { $ref: "#/components/schemas/TaskRef" }
            - { type: "null" }
        subtasks:
          type: array
          items: { $ref: "#/components/schemas/TaskRef" }
        checklist:
          type: array
          items: { $ref: "#/components/schemas/ChecklistItem" }
        comments:
          type: array
          items: { $ref: "#/components/schemas/Comment" }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
        requests:
          description: Newest first (at most 100)
          type: array
          items: { $ref: "#/components/schemas/ChangeRequest" }
    TaskList:
      type: object
      required: [tasks]
      properties:
        tasks:
          type: array
          items: { $ref: "#/components/schemas/TaskSummary" }
    TaskResponse:
      type: object
      required: [task]
      properties:
        task: { $ref: "#/components/schemas/Task" }
    CreateTaskRequest:
      type: object
      required: [title, startDate, endDate]
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/TaskStatus" }
        priority: { $ref: "#/components/schemas/TaskPriority" }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        dueDate: { type: [string, "null"], format: date }
        assigneeIds:
          type: array
          items: { type: string }
        labelIds:
          type: array
          items: { type: string }
        clientId: { type: [string, "null"] }
        parentId: { type: [string, "null"], description: Makes this a subtask (one level deep) }
    UpdateTaskRequest:
      type: object
      description: Only the fields present are changed.
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/TaskStatus" }
        priority: { $ref: "#/components/schemas/TaskPriority" }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        dueDate: { type: [string, "null"], format: date }
        clientId: { type: [string, "null"] }
    UserIdsRequest:
      type: object
      required: [userIds]
      properties:
        userIds:
          type: array
          items: { type: string }
    LabelIdsRequest:
      type: object
      required: [labelIds]
      properties:
        labelIds:
          type: array
          items: { type: string }
...
    ClientRef:
      type: object
      required: [id, name, color]
      properties:
        id: { type: string }
        name: { type: string }
        color: { $ref: "#/components/schemas/LabelColor" }
    TaskRef:
      type: object
      required: [id, title, status]
      properties:
        id: { type: string }
        title: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
    TaskCounts:
      type: object
      required: [subtasks, subtasksDone, checklist, checklistDone, comments, attachments, pendingRequests]
      properties:
        pendingRequests: { type: integer }
        subtasks: { type: integer }
        subtasksDone: { type: integer }
        checklist: { type: integer }
        checklistDone: { type: integer }
        comments: { type: integer }
        attachments: { type: integer }
    ChecklistItem:
      type: object
      required: [id, title, done, assignee]
      properties:
        id: { type: string }
        title: { type: string }
        done: { type: boolean }
        assignee:
          anyOf:
            - { $ref: "#/components/schemas/PersonRef" }
            - { type: "null" }
...
    Comment:
      type: object
      required: [id, body, author, mentions, createdAt]
      properties:
        id: { type: string }
        body: { type: string }
        author:
          anyOf:
            - { $ref: "#/components/schemas/PersonRef" }
            - { type: "null" }
        mentions:
          type: array
          items: { $ref: "#/components/schemas/PersonRef" }
        createdAt: { type: string, format: date-time }
...
    Attachment:
      type: object
      required: [id, filename, mimeType, size, uploadedBy, createdAt, downloadUrl]
      properties:
        id: { type: string }
        filename: { type: string }
        mimeType: { type: string }
        size: { type: integer }
        uploadedBy:
          anyOf:
            - { $ref: "#/components/schemas/PersonRef" }
            - { type: "null" }
        createdAt: { type: string, format: date-time }
        downloadUrl: { type: string }
...
    ChangeRequestKind:
      type: string
      enum: [UPDATE, ASSIGNEE_ADD, ASSIGNEE_REMOVE, SUBTASK_ADD, CHECKLIST_ADD, CHECKLIST_UPDATE, CHECKLIST_REMOVE, ATTACHMENT_REMOVE]
    ChangeRequestStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, CANCELED]
    ChangeRequestPayload:
      type: object
      description: >-
        Flat payload; members by kind — UPDATE: field (status|startDate|endDate|dueDate), value ("" clears dueDate),
        from (server-set); ASSIGNEE_*: userId; SUBTASK_ADD / CHECKLIST_ADD: title;
        CHECKLIST_UPDATE: itemId + title and/or done; CHECKLIST_REMOVE: itemId; ATTACHMENT_REMOVE: attachmentId.
      properties:
        field: { type: string }
        value: { type: string }
        from: { type: string }
        userId: { type: string }
        title: { type: string }
        itemId: { type: string }
        done: { type: boolean }
        attachmentId: { type: string }
    ChangeRequest:
      type: object
      required: [id, kind, status, payload, summary, note, reviewNote, requester, reviewer, createdAt, decidedAt]
      properties:
        id: { type: string }
        kind: { $ref: "#/components/schemas/ChangeRequestKind" }
        status: { $ref: "#/components/schemas/ChangeRequestStatus" }
        payload: { $ref: "#/components/schemas/ChangeRequestPayload" }
        summary: { type: string, description: 'e.g. "Change status to Done"' }
        note: { type: [string, "null"] }
        reviewNote: { type: [string, "null"] }
        requester:
          description: null when the person is no longer a member
          anyOf:
            - { $ref: "#/components/schemas/PersonRef" }
            - { type: "null" }
        reviewer:
          anyOf:
            - { $ref: "#/components/schemas/PersonRef" }
            - { type: "null" }
        createdAt: { type: string, format: date-time }
        decidedAt: { type: [string, "null"], format: date-time }
...
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    Problem:
      type: object
      description: RFC 7807 problem details.
      required: [type, title, status]
      properties:
        type:
          type: string
          description: Machine-readable problem type, e.g. "not_found", "validation".
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"